requirements.txt
marker-pdf>=1.10.0
PyMuPDF>=1.25.0
anthropic>=0.80.0
click>=8.0.0
Pillow>=10.0.0
